Fix: cast Auth::id() to int for SubmissionRepository calls
Auth::id() returns ?int but create()/forUser() expect int. requireLogin()
already guarantees a session, so the cast is safe and silences the
Intelephense type-mismatch diagnostic.

+++ app/controllers/SubmitController.php
@@ -44,7 +44,7 @@ class SubmitController
             return;
         }
 
-        (new SubmissionRepository(Database::pdo()))->create(Auth::id(), $data);
+        (new SubmissionRepository(Database::pdo()))->create((int) Auth::id(), $data);
 
         header('Location: ' . url('/my-submissions'));
         exit;
